refactor testExecutors to base class
